* Merging Delf-Skills, add in NCP exemptions to relevant skills. * Update the temp Piggify Exploit Fix. * Update the temp MassPiggify Exploit fix.

+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillVoidRune.java
@@ -67,9 +67,9 @@ public class SkillVoidRune extends ActiveSkill {
 
         node.set(SkillSetting.DAMAGE.node(), 40);
         node.set(SkillSetting.DURATION.node(), 1500);
-        node.set(SkillSetting.USE_TEXT.node(), ChatColor.GRAY + "["+ChatColor.DARK_GREEN+"Skill"+ ChatColor.GRAY+ "] %hero% imbues his blade with a Rune of "+ChatColor.DARK_PURPLE+"Void.");
-        node.set(SkillSetting.APPLY_TEXT.node(), ChatColor.GRAY + "["+ChatColor.DARK_GREEN+"Skill"+ ChatColor.GRAY+ "] %target% has been silenced by a Rune of Void!");
-        node.set(SkillSetting.EXPIRE_TEXT.node(), ChatColor.GRAY + "["+ChatColor.DARK_GREEN+"Skill"+ ChatColor.GRAY+ "] %target% is no longer silenced!");
+        node.set(SkillSetting.USE_TEXT.node(), ChatColor.GRAY + "[" + ChatColor.DARK_GREEN + "Skill" + ChatColor.GRAY + "] %hero% imbues his blade with a Rune of " + ChatColor.DARK_PURPLE + "Void.");
+        node.set(SkillSetting.APPLY_TEXT.node(), ChatColor.GRAY + "[" + ChatColor.DARK_GREEN + "Skill" + ChatColor.GRAY + "] %target% has been silenced by a Rune of Void!");
+        node.set(SkillSetting.EXPIRE_TEXT.node(), ChatColor.GRAY + "[" + ChatColor.DARK_GREEN + "Skill" + ChatColor.GRAY + "] %target% is no longer silenced!");
         node.set("rune-chat-color", ChatColor.DARK_PURPLE.toString());
 
         return node;
@@ -78,7 +78,7 @@ public class SkillVoidRune extends ActiveSkill {
     @Override
     public String getDescription(Hero hero) {
         int duration = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DURATION, 1500, false) / 1000;
-        int damage = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE, 40, false);
+        double damage = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE, 40, false);
         return getDescription().replace("$1", damage + "").replace("$2", duration + "");
     }
 
@@ -136,8 +136,8 @@ public class SkillVoidRune extends ActiveSkill {
             long duration = SkillConfigManager.getUseSetting(hero, skill, SkillSetting.DURATION, 1500, false);
             double damage = SkillConfigManager.getUseSetting(hero, skill, SkillSetting.DAMAGE, 40, false);
 
-            String applyText = SkillConfigManager.getRaw(skill, SkillSetting.APPLY_TEXT, ChatColor.GRAY + "["+ChatColor.DARK_GREEN+"Skill"+ ChatColor.GRAY+ "] %target% has been silenced by a Rune of Void!").replace("%target%", "$1");
-            String expireText = SkillConfigManager.getRaw(skill, SkillSetting.EXPIRE_TEXT, ChatColor.GRAY + "["+ChatColor.DARK_GREEN+"Skill"+ ChatColor.GRAY+ "] %target% is no longer silenced!").replace("%target%", "$1");
+            String applyText = SkillConfigManager.getRaw(skill, SkillSetting.APPLY_TEXT, ChatColor.GRAY + "[" + ChatColor.DARK_GREEN + "Skill" + ChatColor.GRAY + "] %target% has been silenced by a Rune of Void!").replace("%target%", "$1");
+            String expireText = SkillConfigManager.getRaw(skill, SkillSetting.EXPIRE_TEXT, ChatColor.GRAY + "[" + ChatColor.DARK_GREEN + "Skill" + ChatColor.GRAY + "] %target% is no longer silenced!").replace("%target%", "$1");
 
             // Damage and silence the target
             skill.plugin.getDamageManager().addSpellTarget(targEnt, hero, skill);
